Conflicts:
	src/plugins/websearch/search_snippet.cpp

+++ src/utils/rmd160.h
@@ -48,8 +48,7 @@
 /* adapt these, if necessary, 
    for your operating system and compiler */
 typedef    unsigned char        byte;
-typedef    unsigned long        dword;
-
+typedef    unsigned int        dword;
 
 /********************************************************************/
